+++ Engine/Font.cpp
@@ -53,34 +53,34 @@ void Font::DrawText(std::string text, int x, int y, int size, Color c)
 	int yM = 0;
 	if (costumWidth)
 	{
-		for (int i = 0; i < text.length(); i++)
+		for (int i = 0; i < (int)text.length(); i++)
 		{
 			if (text[i] >= first && text[i] <= last)
 			{
 				float ratio = (float)cRects[(__int64)text[i] - first].GetWidth() / cRects[(__int64)text[i] - first].GetHeight();
-				Vec2 charSize(ratio * size, size);
-				gfx.DrawSurface(RectI(Vei2(x + xM, y), charSize.x, charSize.y), cRects[(__int64)text[i] - first], s, SpriteEffect::ChromaColor(Colors::Magenta, c));
-				xM += (__int64)size * 0.9f * ratio;
+				Vec2 charSize(ratio * size, (float)size);
+				gfx.DrawSurface(RectI(Vei2(x + xM, y), (int)charSize.x, (int)charSize.y), cRects[(__int64)text[i] - first], s, SpriteEffect::ChromaColor(Colors::Magenta, c));
+				xM += (int)((float)size * 0.9f * ratio);
 			}
 			if (text[i] == ' ')
 			{
-				xM += size * 0.7f;
+				xM += (int)(size * 0.7f);
 			}
 		}
 	}
 	else
 	{
 		float radio = 9.f / 13.f;
-		for (int i = 0; i < text.length(); i++)
+		for (int i = 0; i < (int)text.length(); i++)
 		{
 			if (text[i] >= first && text[i] <= last)
 			{
-				gfx.DrawSurface(RectI(Vei2(x + xM, y), size * radio, size), cRects[(__int64)text[i] - first], s, SpriteEffect::ChromaColor(Colors::Magenta, c));
-				xM += size * radio;
+				gfx.DrawSurface(RectI(Vei2(x + xM, y), (int)(size * radio), size), cRects[(__int64)text[i] - first], s, SpriteEffect::ChromaColor(Colors::Magenta, c));
+				xM += (int)(size * radio);
 			}
 			if (text[i] == ' ')
 			{
-				xM += size * radio;
+				xM += (int)(size * radio);
 			}
 		}
 	}
@@ -94,32 +94,32 @@ void Font::DrawTextCentered(std::string text, Vei2 pos, int size, Color c)
 	if (costumWidth)
 	{
 		int width = 0;
-		for (int i = 0; i < text.length(); i++)
+		for (int i = 0; i < (int)(text.length()); i++)
 		{
 			if (text[i] >= first && text[i] <= last)
 			{
 				float ratio = (float)cRects[(__int64)text[i] - first].GetWidth() / cRects[(__int64)text[i] - first].GetHeight();
-				width += (__int64)size * 0.9f * ratio;
+				width += (__int64)(size * 0.9f * ratio);
 			}
 			if (text[i] == ' ')
 			{
-				width += size * 0.7f;
+				width += (int)(size * 0.7f);
 			}
 		}
 		x -= width / 2;
 		y -= size / 2;
-		for (int i = 0; i < text.length(); i++)
+		for (int i = 0; i < (int)text.length(); i++)
 		{
 			if (text[i] >= first && text[i] <= last)
 			{
 				float ratio = (float)cRects[(__int64)text[i] - first].GetWidth() / cRects[(__int64)text[i] - first].GetHeight();
-				Vec2 charSize(ratio * size, size);
-				gfx.DrawSurface(RectI(Vei2(x + xM, y), charSize.x, charSize.y), cRects[(__int64)text[i] - first], s, SpriteEffect::ChromaColor(Colors::Magenta, c));
-				xM += (__int64)size * 0.9f * ratio;
+				Vec2 charSize(ratio * size, (float)size);
+				gfx.DrawSurface(RectI(Vei2(x + xM, y), (int)charSize.x, (int)charSize.y), cRects[(__int64)text[i] - first], s, SpriteEffect::ChromaColor(Colors::Magenta, c));
+				xM += (__int64)(size * 0.9f * ratio);
 			}
 			if (text[i] == ' ')
 			{
-				xM += size * 0.7f;
+				xM += (int)(size * 0.7f);
 			}
 		}
 	}
